Fix overflow for non-mobile modal
In the non-mobile (floating) mode, if the modal content is taller
than the screen the content will overflow but the modal container
won't scroll, causing the content to be cut off.

This fix is a temporary bandage to ensure that overflowed content
is shown. This fix introduces some visual changes to the mobile
modal, which will no longer have its buttons fixed at the
bottom of the screen. But there's no way to fix this for now
without causing more issues. We need to return and fix multiple
other issues with a bigger fix soon.